Portfolio Website
A portfolio manager (PM) is a professional responsible for making investment decisions and carrying out investment activities on behalf of vested individuals or institutions. Clients invest their money into the PM's investment policy for future growth, such as a retirement fund, endowment fund, or education fund.[1] PMs work with a team of analysts and researchers and are responsible for establishing an investment strategy, selecting appropriate investments, and allocating each investment properly towards an investment fund or asset management vehicle.